Refactor tools into partial classes
