+++ RefactorMCP.ConsoleApp/Tools/IntroduceField.cs
@@ -1,0 +1,168 @@
+using ModelContextProtocol.Server;
+using System.ComponentModel;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Formatting;
+using Microsoft.CodeAnalysis.Text;
+
+public static partial class RefactoringTools
+{
+    public static async Task<string> IntroduceField(
+        [Description("Path to the C# file")] string filePath,
+        [Description("Range in format 'startLine:startColumn-endLine:endColumn'")] string selectionRange,
+        [Description("Name for the new field")] string fieldName,
+        [Description("Access modifier (private, public, protected, internal)")] string accessModifier = "private",
+        [Description("Path to the solution file (.sln) - optional for single file mode")] string? solutionPath = null)
+    {
+        try
+        {
+            if (solutionPath != null)
+            {
+                // Solution mode - full semantic analysis
+                var solution = await GetOrLoadSolution(solutionPath);
+                var document = GetDocumentByPath(solution, filePath);
+                if (document != null)
+                    return await IntroduceFieldWithSolution(document, selectionRange, fieldName, accessModifier);
+
+                // Fallback to single file mode when file isn't part of the solution
+                return await IntroduceFieldSingleFile(filePath, selectionRange, fieldName, accessModifier);
+            }
+
+            // Single file mode - direct syntax tree manipulation
+            return await IntroduceFieldSingleFile(filePath, selectionRange, fieldName, accessModifier);
+        }
+        catch (Exception ex)
+        {
+            return $"Error introducing field: {ex.Message}";
+        }
+    }
+
+    private static async Task<string> IntroduceFieldWithSolution(Document document, string selectionRange, string fieldName, string accessModifier)
+    {
+        var sourceText = await document.GetTextAsync();
+        var syntaxRoot = await document.GetSyntaxRootAsync();
+
+        if (!TryParseRange(selectionRange, out var startLine, out var startColumn, out var endLine, out var endColumn))
+            return "Error: Invalid selection range format";
+
+        var startPosition = sourceText.Lines[startLine - 1].Start + startColumn - 1;
+        var endPosition = sourceText.Lines[endLine - 1].Start + endColumn - 1;
+        var span = TextSpan.FromBounds(startPosition, endPosition);
+
+        var selectedExpression = syntaxRoot!.DescendantNodes()
+            .OfType<ExpressionSyntax>()
+            .FirstOrDefault(e => span.Contains(e.Span) || e.Span.Contains(span));
+
+        if (selectedExpression == null)
+            return "Error: Selected code is not a valid expression";
+
+        // Get the semantic model to determine the type
+        var semanticModel = await document.GetSemanticModelAsync();
+        var typeInfo = semanticModel!.GetTypeInfo(selectedExpression);
+        var typeName = typeInfo.Type?.ToDisplayString() ?? "var";
+
+        // Create the field declaration
+        var accessModifierToken = accessModifier.ToLower() switch
+        {
+            "public" => SyntaxFactory.Token(SyntaxKind.PublicKeyword),
+            "protected" => SyntaxFactory.Token(SyntaxKind.ProtectedKeyword),
+            "internal" => SyntaxFactory.Token(SyntaxKind.InternalKeyword),
+            _ => SyntaxFactory.Token(SyntaxKind.PrivateKeyword)
+        };
+
+        var fieldDeclaration = SyntaxFactory.FieldDeclaration(
+            SyntaxFactory.VariableDeclaration(
+                SyntaxFactory.IdentifierName(typeName))
+            .WithVariables(SyntaxFactory.SingletonSeparatedList(
+                SyntaxFactory.VariableDeclarator(fieldName)
+                .WithInitializer(SyntaxFactory.EqualsValueClause(selectedExpression)))))
+            .WithModifiers(SyntaxFactory.TokenList(accessModifierToken));
+
+        // Replace the selected expression with the field reference
+        var fieldReference = SyntaxFactory.IdentifierName(fieldName);
+        var newRoot = syntaxRoot.ReplaceNode(selectedExpression, fieldReference);
+
+        // Add the field to the class
+        var containingClass = selectedExpression.Ancestors().OfType<ClassDeclarationSyntax>().FirstOrDefault();
+        if (containingClass != null)
+        {
+            var updatedClass = containingClass.WithMembers(
+                containingClass.Members.Insert(0, fieldDeclaration));
+            newRoot = newRoot.ReplaceNode(containingClass, updatedClass);
+        }
+
+        var formattedRoot = Formatter.Format(newRoot, document.Project.Solution.Workspace);
+        var newDocument = document.WithSyntaxRoot(formattedRoot);
+        var newText = await newDocument.GetTextAsync();
+        await File.WriteAllTextAsync(document.FilePath!, newText.ToString());
+
+        return $"Successfully introduced {accessModifier} field '{fieldName}' from {selectionRange} in {document.FilePath} (solution mode)";
+    }
+
+    private static async Task<string> IntroduceFieldSingleFile(string filePath, string selectionRange, string fieldName, string accessModifier)
+    {
+        if (!File.Exists(filePath))
+            return $"Error: File {filePath} not found";
+
+        var sourceText = await File.ReadAllTextAsync(filePath);
+        var syntaxTree = CSharpSyntaxTree.ParseText(sourceText);
+        var syntaxRoot = await syntaxTree.GetRootAsync();
+        var textLines = SourceText.From(sourceText).Lines;
+
+        if (!TryParseRange(selectionRange, out var startLine, out var startColumn, out var endLine, out var endColumn))
+            return "Error: Invalid selection range format";
+
+        var startPosition = textLines[startLine - 1].Start + startColumn - 1;
+        var endPosition = textLines[endLine - 1].Start + endColumn - 1;
+        var span = TextSpan.FromBounds(startPosition, endPosition);
+
+        var selectedExpression = syntaxRoot.DescendantNodes()
+            .OfType<ExpressionSyntax>()
+            .FirstOrDefault(e => span.Contains(e.Span) || e.Span.Contains(span));
+
+        if (selectedExpression == null)
+            return "Error: Selected code is not a valid expression";
+
+        // In single file mode, use 'var' for type since we don't have semantic analysis
+        var typeName = "var";
+
+        // Create the field declaration
+        var accessModifierToken = accessModifier.ToLower() switch
+        {
+            "public" => SyntaxFactory.Token(SyntaxKind.PublicKeyword),
+            "protected" => SyntaxFactory.Token(SyntaxKind.ProtectedKeyword),
+            "internal" => SyntaxFactory.Token(SyntaxKind.InternalKeyword),
+            _ => SyntaxFactory.Token(SyntaxKind.PrivateKeyword)
+        };
+
+        var fieldDeclaration = SyntaxFactory.FieldDeclaration(
+            SyntaxFactory.VariableDeclaration(
+                SyntaxFactory.IdentifierName(typeName))
+            .WithVariables(SyntaxFactory.SingletonSeparatedList(
+                SyntaxFactory.VariableDeclarator(fieldName)
+                .WithInitializer(SyntaxFactory.EqualsValueClause(selectedExpression)))))
+            .WithModifiers(SyntaxFactory.TokenList(accessModifierToken));
+
+        // Replace the selected expression with the field reference
+        var fieldReference = SyntaxFactory.IdentifierName(fieldName);
+        var newRoot = syntaxRoot.ReplaceNode(selectedExpression, fieldReference);
+
+        // Add the field to the class
+        var containingClass = selectedExpression.Ancestors().OfType<ClassDeclarationSyntax>().FirstOrDefault();
+        if (containingClass != null)
+        {
+            var updatedClass = containingClass.WithMembers(
+                containingClass.Members.Insert(0, fieldDeclaration));
+            newRoot = newRoot.ReplaceNode(containingClass, updatedClass);
+        }
+
+        // Format and write back to file
+        var workspace = new AdhocWorkspace();
+        var formattedRoot = Formatter.Format(newRoot, workspace);
+        await File.WriteAllTextAsync(filePath, formattedRoot.ToFullString());
+
+        return $"Successfully introduced {accessModifier} field '{fieldName}' from {selectionRange} in {filePath} (single file mode)";
+    }
+
+}
