+++ RefactorMCP.ConsoleApp/Tools/SafeDelete.cs
@@ -1,0 +1,17 @@
+using ModelContextProtocol.Server;
+using System.ComponentModel;
+using Microsoft.CodeAnalysis;
+
+public static partial class RefactoringTools
+{
+    [McpServerTool, Description("Safely delete a field, parameter, or variable with dependency warnings")]
+    public static async Task<string> SafeDelete(
+        [Description("Path to the solution file (.sln)")] string solutionPath,
+        [Description("Path to the C# file")] string filePath,
+        [Description("Line number of the element to delete")] int targetLine,
+        [Description("Type of element (field, parameter, variable)")] string elementType)
+    {
+        // TODO: Implement safe delete refactoring using Roslyn
+        return $"Safe delete {elementType} at line {targetLine} in {filePath} - Implementation in progress";
+    }
+}
